courses/social-network-analysis/homework/001/requirements.txt
matplotlib==3.1.2
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
snap-stanford==5.0.0